Complete Navigation between all screens
A new activity, the main activity, was created. It has a side bar which handles navigation to the schedule and user profile, which are now fragments.

Other options, such as going to the group activity, opening settings,
and logging out are also available on the side panel, altought not all
implemented.