Implement Schedule creation from event/session data
EventAdapter adds a listener where, when the checkbox is checked, the current session is set as completed and a new session is created.

CompareTo functions were added to the FitTime and FitDate DTOs for convenience.

FireStore collection names were moved to constants.

New Event was adjusted to add a first session on top of the created event.